requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.10.5
django-celery-beat==1.0.1
django-celery-results==1.0.1
kombu==4.0.2
Pillow==3.4.2
pytesseract==0.1.6
python-Levenshtein==0.12.0
pytz==2016.10
redis==2.10.5
vine==1.1.3
